refactor: move key from app.rs to key.rs
Key is used by some functions so it should be divided from app.rs
