+++ src/data.rs
@@ -23,7 +23,6 @@ use std::{
     vec::Vec,
 };
 
-use crate::app::{Key, KeyType};
 use crate::parser::{AttributeDefinition, AttributeType, DyneinParser, ParseError};
 use log::{debug, error};
 use rusoto_dynamodb::{
@@ -35,6 +34,7 @@ use tabwriter::TabWriter;
 // use bytes::Bytes;
 
 use super::app;
+use super::key;
 
 /* =================================================
 struct / enum / const
@@ -784,8 +784,8 @@ fn strip_item(
         .collect()
 }
 
-impl From<Key> for AttributeDefinition {
-    fn from(value: Key) -> Self {
+impl From<key::Key> for AttributeDefinition {
+    fn from(value: key::Key) -> Self {
         AttributeDefinition::new(value.name, value.kind)
     }
 }
@@ -800,7 +800,7 @@ fn generate_query_expressions(
     let expression: String = String::from("#DYNEIN_PKNAME = :DYNEIN_PKVAL");
     let mut names = HashMap::<String, String>::new();
     let mut vals = HashMap::<String, AttributeValue>::new();
-    let mut sort_key_of_target_table_or_index: Option<app::Key> = None;
+    let mut sort_key_of_target_table_or_index: Option<key::Key> = None;
 
     match index {
         None =>
@@ -881,12 +881,12 @@ fn generate_query_expressions(
     }
 }
 
-impl From<KeyType> for AttributeType {
-    fn from(value: KeyType) -> Self {
+impl From<key::KeyType> for AttributeType {
+    fn from(value: key::KeyType) -> Self {
         match value {
-            KeyType::S => AttributeType::S,
-            KeyType::N => AttributeType::N,
-            KeyType::B => AttributeType::B,
+            key::KeyType::S => AttributeType::S,
+            key::KeyType::N => AttributeType::N,
+            key::KeyType::B => AttributeType::B,
         }
     }
 }
@@ -894,7 +894,7 @@ impl From<KeyType> for AttributeType {
 /// Using existing key condition expr (e.g. "myId <= :idVal") and supplementary mappings (expression_attribute_names, expression_attribute_values),
 /// this method returns GeneratedQueryParams struct. Note that it's called only when sort key expression (ske) exists.
 fn append_sort_key_expression(
-    sort_key: Option<app::Key>,
+    sort_key: Option<key::Key>,
     partition_key_expression: &str,
     sort_key_expression: &str,
     mut names: HashMap<String, String>,
